Migrate to typescipt (#2)
* refactor: migrate to typescript

* feat: add the clean script
