+++ package.json
@@ -5,6 +5,8 @@
     "main": "index.js",
     "type": "module",
     "scripts": {
+        "clean": "rimraf dist",
+        "build": "tsc",
         "start": "node ./index.js",
         "start:dev": "nodemon ./index.js",
         "test": "echo \"Error: no test specified\" && exit 1"
@@ -13,9 +15,12 @@
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/node": "^22.8.0",
+        "nodemon": "^2.0.16",
+        "typescript": "^5.6.3"
     },
     "dependencies": {
+        "rimraf": "^6.0.1",
         "ws": "^8.8.0"
     }
 }
